Refine some code to make code run safely.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@10873 6f19259b-4bc3-4df7-8a09-765794883524

+++ EdkCompatibilityPkg/Compatibility/FrameworkHiiOnUefiHiiThunk/OpcodeCreation.c
@@ -717,7 +717,7 @@ F2UCreateNumericOpCode (
   // We need to create a default value.
   //
   ZeroMem (&UOpcodeDefault, sizeof (UOpcodeDefault));
-  UOpcodeDefault.Header.Length = sizeof (UOpcodeDefault);
+  UOpcodeDefault.Header.Length = (UINT8) sizeof (UOpcodeDefault);
   UOpcodeDefault.Header.OpCode = EFI_IFR_DEFAULT_OP;
 
   UOpcodeDefault.DefaultId = 0;
